Functional AI! MVP!
However, I do see a few problems with syntax.  The purpose of this exercise was to understand how to completely write my AI chat bot in a text editor file saved in a .html format.  I also see that no matter what I tried before, the sentences can still go on and on, even though I placed a limit on the box size.  I also noticed that both the answers and queries remain, and not in a listed view, at the very least.  Much growth and improvement is to be desired, but I did it! I have a minimum viable product (MVP)! Thank you, JESUS🙌🏿👏🏿